requirements_tmp.txt
python-dotenv==1.0.1
numpy~=1.26.4
pandas==2.2.2
librosa==0.10.2.post1
xgboost==2.0.3
pydub==0.25.1
gammatone==1.0.1
scipy~=1.13.1
spafe==0.3.3